Move helper functions to actions class
